Basic UI Functionality
Without the new designs added, the Pause Menu is now fully functional with an Options Screen. Additionally, a checkpoint popup that displays unique text depending on the checkpoint now exists, and a trigger box for it. Options Menu stuff has been implimented but is quite broken at the current moment.